requirements.txt
# Core
torch>=2.2.0
torchaudio>=2.2.0
numpy
scipy

# HuggingFace
datasets>=4.0.0
transformers>=4.40.0
huggingface-hub
soundfile
librosa

# MT Evaluation
unbabel-comet>=2.2.0

# SONAR / BLASER
sonar-space>=0.4.0

# Ensemble
lightgbm

# Utilities
pandas
matplotlib
seaborn
scikit-learn
tqdm
